rensar upp

+++ src/main/java/com/temp/webshop/webshop/repository/HistoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface HistoryRepository extends JpaRepository<History, Integer> {
+public interface HistoryRepository extends JpaRepository<History, Long> {
     public List<History> findPurchaseHistoryByUserId(Long id);
 }
